refactor(shared): own overlay-app + detail-extension registration surfaces (#12093 item 6)
Decouple the Node app-core barrel from the React @elizaos/ui package's
internal file layout for the app-registration surfaces. The overlay-app and
detail-extension registries — plus the isomorphic helpers they need
(registry-host store, AOSP user-agent detection) — were owned by @elizaos/ui
and re-exported into Node-reachable app-core modules by exact deep ui path,
so any ui file move broke app-core and the API barrel pulled the React graph.

Move the single canonical implementation into @elizaos/shared:
- packages/shared/src/apps/{overlay-app-api,overlay-app-registry,
  detail-extension-registry,detail-extension-types}.ts
- packages/shared/src/registry-host.ts (UI registry host store singleton)
- packages/shared/src/platform/aosp-user-agent.ts

The former @elizaos/ui files become thin re-export shims (matching the
existing errors.ts / serialise.ts sibling pattern), so the large ui consumer
graph is unchanged. app-core's Node-reachable ui-compat.ts + browser.ts now
source registerOverlayApp / registerDetailExtension and the OverlayApp /
AppDetailExtensionProps contracts from @elizaos/shared instead of the React
package.

Seam test (packages/shared/src/apps/registry.test.ts) owns the canonical
AOSP-gating edge cases the ui copy previously guarded; the ui shim test now
exercises the shim -> shared path, proving both packages resolve one
implementation.

Refs #12093
