+++ packages/app-core/src/ui-compat.ts
@@ -1,5 +1,19 @@
+// Registration-surface contracts + registries (overlay apps, detail extensions)
+// are owned by @elizaos/shared — the React-free canonical home — so this
+// Node-reachable shim registers app surfaces without touching the React package.
 export type {
   AppDetailExtensionProps,
+  OverlayApp,
+  OverlayAppContext,
+} from "@elizaos/shared";
+export { registerDetailExtension, registerOverlayApp } from "@elizaos/shared";
+// Everything below re-exports from its narrow `@elizaos/ui` subpath rather than
+// the root barrel. The barrel (`@elizaos/ui`) eagerly evaluates the entire
+// frontend component graph, and this shim is reachable from the Node
+// `@elizaos/app-core` barrel (index.ts) — so importing it from the bare barrel
+// dragged ~1000 React modules (and their deps) into the API process at boot.
+// Subpath imports pull only the specific component. Mirrors `browser.ts`.
+export type {
   AppRunSummary,
   AppSessionJsonValue,
   FeedActivityItem,
@@ -9,19 +23,9 @@ export type {
   FeedPredictionMarket,
   FeedTeamAgent,
   FeedWallet,
-  OverlayApp,
-  OverlayAppContext,
   SurfaceTone,
 } from "@elizaos/ui";
-// Re-export each value from its narrow `@elizaos/ui` subpath rather than the
-// root barrel. The barrel (`@elizaos/ui`) eagerly evaluates the entire frontend
-// component graph, and this shim is reachable from the Node `@elizaos/app-core`
-// barrel (index.ts) — so importing it from the bare barrel dragged ~1000 React
-// modules (and their deps) into the API process at boot. Subpath imports pull
-// only the specific component. Mirrors `browser.ts`. The `export type` block
-// above is erased at compile time and needs no narrowing.
 export { client } from "@elizaos/ui/api";
-export { registerDetailExtension } from "@elizaos/ui/components/apps/extensions/registry";
 export {
   SurfaceBadge,
   SurfaceCard,
@@ -36,7 +40,6 @@ export {
   toneForStatusText,
   toneForViewerAttachment,
 } from "@elizaos/ui/components/apps/extensions/surface.helpers";
-export { registerOverlayApp } from "@elizaos/ui/components/apps/overlay-app-registry";
 export { PagePanel } from "@elizaos/ui/components/composites/page-panel";
 export { Button } from "@elizaos/ui/components/ui/button";
 export { Input } from "@elizaos/ui/components/ui/input";
