+++ packages/shared/src/apps/detail-extension-types.ts
@@ -1,0 +1,9 @@
+import type { ComponentType } from "react";
+import type { RegistryAppInfo } from "../contracts/apps.js";
+
+export interface AppDetailExtensionProps {
+  app: RegistryAppInfo;
+}
+
+export type AppDetailExtensionComponent =
+  ComponentType<AppDetailExtensionProps>;
