+++ packages/shared/src/apps/index.ts
@@ -1,0 +1,4 @@
+export * from "./detail-extension-registry.js";
+export * from "./detail-extension-types.js";
+export * from "./overlay-app-api.js";
+export * from "./overlay-app-registry.js";
