+++ packages/ui/src/components/apps/extensions/types.ts
@@ -1,9 +1,5 @@
-import type { ComponentType } from "react";
-import type { RegistryAppInfo } from "../../../api";
-
-export interface AppDetailExtensionProps {
-  app: RegistryAppInfo;
-}
-
-export type AppDetailExtensionComponent =
-  ComponentType<AppDetailExtensionProps>;
+// Re-export shim: detail-extension types now live in `@elizaos/shared`.
+export type {
+  AppDetailExtensionComponent,
+  AppDetailExtensionProps,
+} from "@elizaos/shared";
