+++ packages/ui/src/registry-host.ts
@@ -1,29 +1,9 @@
-export interface UiRegistryHost {
-  getStore<T>(key: string, create: () => T): T;
-}
-
-class DefaultUiRegistryHost implements UiRegistryHost {
-  private readonly stores = new Map<string, unknown>();
-
-  getStore<T>(key: string, create: () => T): T {
-    const existing = this.stores.get(key);
-    if (existing !== undefined) return existing as T;
-    const created = create();
-    this.stores.set(key, created);
-    return created;
-  }
-}
-
-let activeRegistryHost: UiRegistryHost = new DefaultUiRegistryHost();
-
-export function provideUiRegistryHost(host: UiRegistryHost): void {
-  activeRegistryHost = host;
-}
-
-export function getUiRegistryStore<T>(key: string, create: () => T): T {
-  return activeRegistryHost.getStore(key, create);
-}
-
-export function resetUiRegistryHostForTests(): void {
-  activeRegistryHost = new DefaultUiRegistryHost();
-}
+// Re-export shim: the UI registry host now lives in `@elizaos/shared` so Node
+// app-registration code shares one canonical store singleton without importing
+// the React package. See `@elizaos/shared/src/registry-host.ts`.
+export {
+  getUiRegistryStore,
+  provideUiRegistryHost,
+  resetUiRegistryHostForTests,
+  type UiRegistryHost,
+} from "@elizaos/shared";
